[refactor]: move current fss impl to continuous
